put both sketches in the same file
