+++ crates/polars-expr/src/reduce/approx_quantile.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use polars_compute::req::ReqSketchBounded;
+use polars_compute::approx_quantile::ReqSketchBounded;
 use polars_core::with_match_physical_numeric_polars_type;
 use polars_utils::total_ord::TotalOrd;
 use rand::SeedableRng;
