+++ crates/polars-ops/src/series/ops/approx_quantile.rs
@@ -1,10 +1,10 @@
-use polars_compute::req;
+use polars_compute::approx_quantile::ReqSketchBounded;
 use polars_core::prelude::*;
 
 pub fn approx_quantile(s: &Column, quantile: &Series, error: f64) -> PolarsResult<Scalar> {
     let s = s.as_materialized_series();
     let ca: &Float64Chunked = s.as_ref().as_ref();
-    let mut sketch = req::ReqSketchBounded::new(error, 1000_000);
+    let mut sketch = ReqSketchBounded::new(error, 1000_000);
     for item in ca.iter() {
         if let Some(item) = item {
             sketch.update(&[item]);
